Add support for verbatim text.
Code between `verbatim` and `endverbatim` is copied (almost) verbatim to
the output. The exception is that color names can be used inside
verbatim code, preceded by an @ sign. They are interpolated and replaced
by their values. For example, this code inside a template:

    verbatim
    if get(g:, 'fancyflag', 0)
      hi! Normal ctermfg=@red ctermbg=@black cterm=NONE,bold guifg=@brightred gui=NONE,undercurl guisp=@yellow 
    endif
    endverbatim

may be translated into something like:

    if get(g:, 'fancyflag', 0)
      hi! Normal ctermfg=88 ctermbg=0 cterm=NONE,bold guifg=#ff6400 gui=NONE,undercurl guisp=#ffff00
    endif

in the output.

Code between `verbatim` and `endverbatim` is highlighted as Vim syntax.