Update Gruvbox template.
Let Colortemplate infer the best approximate colors for the terminal,
with the exception of red for the dark background (Colortemplate's
choice is a bit too bright).

There are several differences compared to the values in the original
version of Gruvbox. IMHO, Colortemplate's values are better, especially
for the light background, which is much less yellowish.