Remove support for alternative colors for transparent background.
Treating the option for setting a transparent background in a special
way is a bit artificial and limiting. This commit removes support for
that feature. Instead, it is the user's responsibility to add such
support.

This simplifies Colortemplate's internal structures a bit and solves
potential a few issues. For example, before this commit the following
would result in a wrong colorscheme:

verbatim
if g:foo
endverbatim
MyHiGroup black white/none
verbatim
else
endverbatim
MyHiGroup red green/none
endverbatim

because the transparent versions of MyHiGroup would be moved to the
beginning of the colorscheme's definition into a conditional block (and
there would be two different MyHiGroup definitions further below).