Update default clone template.
This is exactly like the default colorscheme except for the fact that
the clone defines the Normal group (in default it's cleared), and the
Cursor group in the terminal (in default Cursor exists only in GUI Vim).

I have added color white231 because, in the dark variant, white in the
terminal is color 231 instead of 15 (white in the GUI is always 15).