Fix colorschemes not being saved when already opened in a buffer.
Fix an issue that prevented colorschemes to be saved when they are
already opened in another buffer, without Colortemplate issuing any
error.

I thought that silent! would force writing the buffer in those
situations, but that is not the case. Instead, now the script checks
whether the path to be written is already opened in another buffer. If
it is, raise an error by default. If the user forces overwriting data
with Colortemplate!, delete the buffer so that the new one can be saved.