Fallback to 16 colors transparently when t_Co < 256.
When Terminal Colors is set to `256,16`, 256 color palette is preferred
and 16 color palette may be requested by the user by setting *_use_16
option.

Before this commit, when t_Co was less than 256 and *_use_16 was not set
by the user, the colorscheme would show the error 'There are not enough
colors'. This commit implements automatic fallback to 16 colors in that
case, instead. The choice can still be overridden by the user by
setting *_use_16 explicitly.

Fixes https://github.com/lifepillar/vim-colortemplate/issues/9.