Output contrast ratio matrix by default.
After compiling a template with :Colortemplate, one or two buffers
(depending on the number of backgrounds) are opened showing a contrast
ratio matrix for the corresponding background. The matrix shows the
contrast ratio between all pairs of colors. Some text above the matrix
helps interpreting the numbers.

To suppress the contrast ratio buffers, set
g:colortemplate#constrast_matrix to 0.