Thorough rewrite. First commit heading to v2.
Lots of changes here. This started as an attempt to simplify the code,
but the result is messier than before. So be it, VimScript is not the
most elegant language anyway.

Among the most important changes:

- `Terminal Colors` has been deprecated and does nothing; there is a new
  `Variant` keyword now, which allows you to choose the active variants
  to which subsequent definitions apply. Possible values are `gui` or a
  number (e.g., 256, 16, 8, 2).

- `Background` directives (and `Variant`) directives may appear
  anywhere in the code and may be repeated many times. `Background` has
  a new `any` value, which makes the subsequent definitions valid for
  both dark and light background.

- Support for italics is detected automatically and the colorscheme
  adapts accordingly.

- The generated colorscheme now keeps definitions for each variant
  separate: this should make the colorscheme load even faster than
  before in the common cases (GUI and 256-color terminals).

- Etc etc... See the code.

Some things are currently broken or not working, such as aux files,
stats and some validations. The broken functionality will be restored in
the next commits.

Tests are broken, too. They will be fixed later on.