Generate code to fix the "ctermbg=234" bug.
The "ctermbg=234" (from the example originally used to reproduce it) is
a bug in Vim < 8.0.0616, causing the background to be switched from dark
to light or vice versa after executing `:hi Normal ...`.

The bug happens when:

- `background` is `dark` and `hi Normal ctermbg=NNN` is executed, with
  `NNN` equal to `7`, `9`,  or in the range`10`–`255`; or

- `background` is `light` and `hi Normal ctermbg=NNN` is executed, `NNN`
   equal to `8` or in the range `0`–`6`.

This commit adds code to check whether such conditions are satisfied and
automatically generates code to address the issue (the problem is
corrected by setting `background` again after `hi Normal`). 

Fixes https://github.com/lifepillar/vim-colortemplate/issues/13.