Simplify directory handling in the toolbar.
The toolbar is window-specific, but the working directory is a
buffer-local variable. This was a source of mismatches before this
commit, e.g., when switching buffers.

The working directory is set to the directory of the current buffer (or
to its parent directory if the current buffer is inside a `templates`
directory) and this is generally what the user expects. So, simply show
a Dir menu in the toolbar, there is no need to show the name of the
directory. Clicking on Dir still shows the user what the working
directory is and gives a chance to change the path.