Add the output directory to runtimepath by default.
This ensures that the correct color scheme (yours!) is found when more
color schemes with the same name exist. For instance, if you call your
color scheme `blue` (like one of Vim's color schemes) and the value of
b:colortemplate_outdir is not prepended to `runtimepath`, then when you
execute `:colorscheme blue`, chances are good that Vim's color scheme
will be loaded instead of your color scheme.

This also ensures that your color scheme is found when its contained in
a directory that is not already in `runtimepath` or `packpath`. So, you
may keep the color schemes under development outside your ~/.vim folder,
say.

A new option is added to turn off this behaviour, i.e., to leave
`runtimepath` untouched.