Accept `none` as a base-16 color value.
With this commit, a base-16 color definition may use the special keyword
`none` (all lowercase) to mean that in <=16 color schemes the
cterm(fg|bg) values must be set to NONE.

For instance, this:

    Color: foo #ffffff #000000 ~ none
    Variant: 16
    Normal foo omit

is translated into

    hi Normal ctermfg=NONE

Note that this:

    Color: foo #ffffff #000000 ~
    Variant: 16
    Normal foo omit

still gives an error, because the base-16 color is undefined.

**TODO:** proper syntax highlighting.

Fixes https://github.com/lifepillar/vim-colortemplate/issues/24.