(Tentatively) add `;` for line comments.
As the subject line says, this is *tentative*. I have opened a Github
issue for discussion:

    https://github.com/lifepillar/vim-colortemplate/issues/30

The feature is undocumented and `;` should not be used until I settle on
that symbol.

This commit is backward compatible: it just adds `;` as a new symbol for
line comments, but `#` can still be used.