Use a popup window to dynamically update styles.
Instead of defining a lot of global mappings, define a command that
opens a popup window and allows to toggle the styling attributes (bold,
italics, etc.) of the highlight group under the cursor.

This is just a draft implementation. It works, but it can be improved in
many ways.

Requires support for popup window in Vim, of course.