Fix bug causing spurious @suppress statements.
Fix a bug that caused Colortemplate to generate spurious @suppress
statements when g:colortemplate_ignore_missing was set.

For instance, with reference to the committed test file, before this
commit the output would contain the following lines:

    @suppress PopupSelected
    @suppress QuickFixLine

despite the fact that such highlight groups are defined in the source as
linked groups.