Fix base-256 color 0 being skipped.
When defining a base 256 color as 0 (code for terminal black) with Color
and then setting a highlight to it, it gets ignored. For instance:

    Color: black #000000 0
    EndOfBuffer black none

generates

    hi EndOfBuffer ctermbg=NONE cterm=NONE

instead of

    hi EndOfBuffer ctermfg=0 ctermbg=NONE cterm=NONE

This is due to a bug in `s:hi_item()`. In particular, the culprit is the
somewhat surprising fact that

    0 ==# 'non-empty string'

evaluates to true in Vim. Tests of such kind are performed by
`s:hi_item()`.

This commit fixes the issue by explicitly checking whether the `value`
argument passed to `s:hi_item()` is a String. Interestingly, trying to
cast `value` to String instead (as in: `string(a:value) ==# 'omit')
breaks most of the tests.

I am not entirely satisfied by this fix, although it works (partially,
but not exclusively, because I do not fully understand why the above
mentioned casting breaks havoc). A better approach would consist in
revising the internal data structure that stores the color definitions,
where some fields can host values of different types (Numbers or
Strings). It would probably be better to store everything only as
a String, and cast to numbers only if necessary.

Fixes: https://github.com/lifepillar/vim-colortemplate/issues/36