Inherit unspecified base-16 color from base-256 value less than 16.
When the base-16 value is not specified and the base-256 value is less
than 16, set the base-16 value to that value.

Note that, if the base-256 value is not less than 16, then the base-16
value will remain unspecified (i.e., -1).

This avoids a "Base-16 value undefined" error in check_bug_bg234(). The
latter could probably be re-written to avoid the error without the
changes introduced by this commit.