Officially deprecate # for comments.
The # symbols for comments has been deprecated to avoid confusion with
its use in commands (#if, #call, etc.).

🎺 The new symbol for comments is a semicolon (;) 🎺

Colortemplate now raises a warning when the deprecated symbol is used.
Upgrading a color scheme, however, shouldn't be much more difficult
than:

    :%s/# /;

The reason why I have preferred ; to -- (which was the other candidate):

- easier to parse;
- easier to type;
- cannot be confused with anything in Colortemplate's syntax (while - is
  used also in -> for linked groups).

Resolves https://github.com/lifepillar/vim-colortemplate/issues/30.