Generate a scaffold from the current color scheme.
Define a function colortemplate#import#run() that processes the
currently active highlight group definitions and generates a template.

NOTE: this is NOT meant to be a full-fledged translation from Vim script
back to Colortemplate! The goal is only to facilitate the initial
scaffolding of a template.

This commit was prompted by this feature request:
https://github.com/lifepillar/vim-colortemplate/issues/32