Allow setting a step for the sliders in the style popup.
Just select a slider then type a number between 1 and 99: That will
become the new step for all sliders. This way colors can be changed more
quickly.

When the marker is not on a slider, digits are passed to the underlying
buffer. This allows you to continue using counts for mappings (e.g., to
type 5j to move down five lines).