Provide better error message when hi group definition is ambiguous.
Colortemplate does not accept the following:

Background: dark
Color: foo #000000 ~
Background: light
Color: foo #ffffff ~
Background: any
Normal foo foo

Although that might seem unintuitive, it makes sense once you think that `Background: any` is for code that must be put *before* checking the background (so, `Background: any` must be interpreted as “What comes below does not depend on which background is active”).

Before this commit, Colortemplate would generate an “Undefined color” error, which is not very informative. With this commit, the error message clearly explains what the error is, and how to fix it.