Support `Variant: 0` (t_Co<2). Use cterm attributes when t_Co=2.
The lowest supported variant is currently 2 based on the assumption that
cterm* are ignored in favor of term when &t_Co <= 2. This is wrong
because cterm* are definitely used when &t_Co == 2 (Romain Lafourcade).

This commit fixes the above issue by changing Colortemplate's generated
code as follows:

- when t_Co = 2, use cterm attributes;
- when t_Co < 2 or t_Co is undefined or t_Co does not exist, use term
  attributes.

TODO: when t_Co < 8, should we require that base-16 colors do not
exceed t_Co?