Import term and cterm attributes
Import, parse and generate term and cterm attributes separately.

The `cterm` attributes are imported as `term` in the "Variant: gui 256".
Import as common attributes if `gui` and `cterm` attributes are the
same, per specific attribute otherwise.

Import term attributes in Variant 0.
A `Variant: 0` is always created on import.
The common attributes in this variant are the imported "term=" attributes.