Import and generate term and cterm attributes
Import, parse and generate term and cterm attributes separately.

Import as common attributes if all of `gui`, `cterm`, and `term`
attributes are the same, per specific attribute otherwise.

Parse `c` as short name for the `cterm` attribute.

Generate a "no_italics" overwrite with the `term` attribute only for
monochrome terminals.

Add `cterm` as known attribute in the colortemplate syntax file.

Adapt tests 22, 70, and 84. Keep the expected results as is.