[relalg] Add tentative generalization for Divide.
Implement a division operator that works with any pairs of relations.

R(X,Y) div S(Y,Z), with X and Z disjoint, returns a relation Q(X,Z) with
the following property: for every (x,z) in Q(X,Z), joining (x,z) with
any tuple of S and projecting on (X,Y) produces a tuple in R.

This operator reduces to standard division when the schema of S is a
subset of the schema of R—except for some corner case (when S is empty).

The implementation is obviously not optmized yet.