[relalg] Add SumBy, CountBy, MaxBy, MinBy, AvgBy.
These are more efficient and concise than using GroupBy.

Note that these are leaf operators returning a relation—not a scalar.
The schema of the output relation consists of the group-by attributes
plus an attribute for the computed aggregate values.

The corner case of an empty group-by set of attributes has to be fixed.