[relalg] Let SumBy and CountBy return 0 on empty input.
This commit fixes the result of counting of summing over an empty
relation. The rationale is as follows:

- If `groupBy` is empty, all the tuples must end up in the same group
  (because they all coincide on the values of all the `groupBy`
  attributes, trivially). If the relation is empty, this group will be
  empty. Then, the sum over the elements of this group is zero, and the
  count of the elements of the group is zero.
- If `groupBy` is not empty, a group is created for each distinct value
  of the `groupBy` attributes. If the relation is empty, there are zero
  distinct values, hence zero groups. Therefore, there is nothing to sum
  or count over. The result is an empty relation.

In case you are not convinced that in the latter case the empty relation
is the correct output, please note that, by definition, these operators
must output a relation over a schema consisting of the `groupBy`
attributes plus an additional attribute for the aggregate. In the second
case, no tuple can be returned because there are no values to assign to
the `groupBy` attributes.

Note that, for the other operators (Min, Max, Avg), if the input
relation is empty, the output is always the empty relation. Reasoning as
above, it should not be difficult to see that this is correct when the
`groupBy` argument is not empty. But when `groupBy` is empty, one would
expect the output to contain exactly one tuple: {min: …}, {max: …},
{avg: …}, respectively. The problem is that Min, Max, and Avg are not
defined when the input is the empty set. So, there is no value to assign
to such tuples. An "inapplicable" marker should be used: the obvious
candidates are `v:none` and `null`. I do not want to use either until
all the implications are clear to me (`null` in Vim is *not* like SQL's
NULL, but still it may introduce some logical complications—`v:none`
has a special type, and this may cause unwanted or surprising type
mismatch errors).