[relalg] Access private members in lambda.
Vim before patch 9.0.1240 did not allow accessing private members in a
lambda defined inside a class. So, a few accessors were defined to allow
Delete() to access indexes and constraints. But now those can be
removed.

I leave Index(), for now, because is used in a test. I may decide to
remove it in the future and put the test in librelalg.vim.