[parser] Pass context/state to Map and Apply.
This commit enables the user to maintain arbitrary parser's state inside
the context.

I have tried different solutions for managing the parser's state, and
this commit is the result of that effort, which takes into account the
current limitations (read: bugs) in Vim 9 script.

Using inheritance would be a simple and elegant alternative to the
current code, but I gave up on that due to a few bugs in Vim (at least
up to 9.0.1321), in particular https://github.com/vim/vim/issues/12023.
See the commented out code in the test file.

I have also tried decoupling context and state using interfaces, but
that did not buy me much (except complexity).

All in all, the current solution is even simpler than using inheritance,
and requires less boilerplate code (no need to subclass `Context`).