[relalg] Fix bottleneck created by indexes.
The hierarchical implementation of indexes is extremely slow. Inserting
a few thousands tuples takes (tens of) seconds, while the elapsed time
should be of the same order of magnitude as for plain list addition
(tens of milliseconds for a few thousands calls to add()).

Therefore, use a flat hash for indexing. This brings times within
reasonable expectations. Clashing of dictionary keys should not be
possible.