[relalg] Perform EquiJoin in linear time in some cases.
When the second argument of an EquiJoin is a Rel and the join is on a
key, use lookup to perform the join in linear time.

A further improvement, left for a future commit, would be to generate an
index on the fly. This would speed up joins on derived relations, too.