Add experimental generator.
Instead of generating raw highlight group commands, this generator
imports a class with predefined methods to generate those commands on
the fly. To use it, change the import in colortemplate.vim from

    import './generator.vim' as generator

to

    import './generator_experimental.vim' as generator

I wrote this for two reasons: first, to test how easy it is to develop a
new generator (it literally took me less than 10 minutes to write this
proof of complete—ok, it's very incomplete, but still...). Second, to
check the performance of Vim 9 script when generating definitions in
this way.

A simple test with dark.colortemplate results in loading times of ~1.2ms
for the default generator vs ~3ms for this experimental generator. This
is not surprising, as most of the time is spent in `execute` commands,
and I believe that the performance of `execute` has not changed much in
Vim 9 script compared to legacy Vim script.