Do not omit special color.
Omitting the special color by default may have unexpected consequences.
For instance, if you define the special color for a highlight group in
the default definition, and then you override the definition without
specifying a special color, then the special color of the default
definition will apply also to the overridden definition.