Set s:t_Co in a way that pleases more Neovim.
Set s:t_Co in a way that pleases more Neovim when Neovim support is
required.

Neovim has deprecated all `t_xx` options including `t_Co`, which is
currently read-only (but could potentially be removed). According to
Neovim's developers, if a user wants to force the use of 16 colors even
in terminal that supports more colors, it's the colorscheme that must do
so. See in particular this comment:

    https://github.com/neovim/neovim/issues/11974#issuecomment-593795588

So, this PR defines `s:t_Co` prioritizing the existence of a colorscheme
specific variable call `g:<prefix>_t_Co` or of a global `g:t_Co`
setting, falling back to checking whether `t_Co` exists otherwise.