Generate gui attributes for overriding definitions.
When a terminal variant applies, overriding definitions should consider
that termguicolors may be set. So, it is not enough to generate cterm
attributes for them, but gui attributes must be generated, too.

That is not necessary for the default definitions, because gui
attributes are always applied by default.

Failing to explicitly override gui attributes resulted in issues such as
this: https://github.com/lifepillar/vim-solarized8/issues/105.