[reactive] Leave pool management to the user.
Adding pools to the library was a mistake: it complicates the library
code, but it also makes memory management significantly more complex for
the user.

The original and only purpose of pools was to make it seamless to clear
the effects of all properties. Therefore, a better solution is to just
accept an optional list of properties when instantiating a property, and
have the constructor add the new property to the list. Then, the client
is able to easily keep track of all the properties it creates.