[reactive] Move property initialization to an Init() method.
This is to make it easier for clients to specialize (subclass) Property.
As explained in `:help extends`, the constructor of the base class
cannot be invoked, and it is not invoked automatically. For a child
class to be able to do the same initialization as the parent class, the
initialization code should be moved to another method—which is what this
commit does.