Redesign the data model
Use a simpler data model. This removes the constraint that default highlight
group definitions must appear before overrides.

Also move most of the autoload script into `import`.

This is WIP: the data model and the parser are ok (barring bugs), but the
generator, the color statistics and the importer need to be updated.