Put color scheme path in 'runtimepath'.
This basically reverts 386d71ab9a806d4d3db21c73d92155d15966f594.
The path of a color scheme must be added to `runtimepath` for the color
scheme to be found also by external commands. For example, a user may
try to load the color scheme directly with `:colorscheme`. Or another
plugin may do so (StylePicker, for example).

The `colortemplate_rtp` option is not restored because not setting 'runtimepath'
may cause anomalous behaviour.

Signed-off-by: Lifepillar <lifepillar@lifepillar.me>