Generate code that supports termguicolors consistently.
The idea of these changes is to consolidate the semantics
of a template while accounting for `termguicolors`.

When `termguicolors` is set, Vim uses `gui[fg|bg|sp]` attributes
in the terminal, but keep using `cterm` attributes (because
termgui*colors* is, well, about colors).

Now a template must be interpreted in the same way:  `/gui`
overrides apply to terminals with `termguicolors` set, but limited
only to the color attributes. The style attributes are always taken
from whatever terminal definition is active. Any deviation from this
rule should be considered a bug.

For instance:

```
Environments: gui 256
Title orange lightgreen italic
 /256 red green bold
```

With this definition, a true-color terminal with `termguicolors` set
will display `Title`s with an orange foreground, a lightgreen background,
**but in boldface.**

If you want to use the GUI attributes for true-color terminals, you must
explicitly do so. A predefined `tgc` discriminator  can be used for the
purpose. For example:

```
Environments: gui 256
Title orange lightgreen italic
 /256 red green bold
/256+tgc true omit omit italic
```

See `test316` and `test317` for other examples.

Signed-off-by: Lifepillar <lifepillar@lifepillar.me>