Allow different attributes for GUI and true-color terminals.
By default, true-color terminals with `termguicolors` on adopt
whatever colors (`guifg`/`guibg`) have been defined for the
GUI (via a default definition or a `/gui` override) and whatever
terminal attributes are active (via a default definition or a
`/N` override).

In rare circumstances, one may want to use different colors
in a true-color terminal than in GUI Vim (for instance, one may
want normal foreground with a red undercurl for SpellBad in
the GUI, but a red foreground with underline in the terminal).

Users may expect that using the `+tgc` discriminator would
achieve their goal, but before this commit it didn't.

Fixes https://github.com/lifepillar/vim-colortemplate/issues/97.

Signed-off-by: Lifepillar <lifepillar@lifepillar.me>